fix bug ang phan quyen he thong

+++ app/Http/Controllers/AccountLevelController.php
@@ -2,8 +2,13 @@
 
 namespace App\Http\Controllers;
 
+use App\Constants\ActionConstant;
+use App\Constants\PermissionConstant;
+use App\Exceptions\ForbiddenException;
 use App\Services\AccountLevelService;
 use App\Services\BaseService;
+use App\Trait\RoleAndPermissionTrait;
+use Exception;
 use Illuminate\Http\JsonResponse;
 use Illuminate\Http\Request;
 
@@ -17,6 +22,7 @@ class AccountLevelController extends Controller
 
     public BaseService $service;
 
+    use RoleAndPermissionTrait;
     public function __construct()
     {
         $this->service = new AccountLevelService();
@@ -24,6 +30,9 @@ class AccountLevelController extends Controller
 
     public function index(): JsonResponse
     {
+        if (!$this->hasPermission(PermissionConstant::account_level(ActionConstant::LIST)))
+            throw new ForbiddenException(__('Access denied'), new Exception());
+
         return response()->json($this->service->getAllByAdmin());
     }
 
@@ -50,6 +59,9 @@ class AccountLevelController extends Controller
      */
     public function store(Request $request): JsonResponse
     {
+        if (!$this->hasPermission(PermissionConstant::account_level(ActionConstant::ADD)))
+            throw new ForbiddenException(__('Access denied'), new Exception());
+
         return response()->json($this->service->add($request));
     }
 
@@ -84,6 +96,9 @@ class AccountLevelController extends Controller
      */
     public function update(Request $request, $id): JsonResponse
     {
+        if (!$this->hasPermission(PermissionConstant::account_level(ActionConstant::EDIT)))
+            throw new ForbiddenException(__('Access denied'), new Exception());
+
         return response()->json($this->service->update($id, $request));
     }
 
@@ -95,6 +110,9 @@ class AccountLevelController extends Controller
      */
     public function destroy($id): JsonResponse
     {
+        if (!$this->hasPermission(PermissionConstant::account_level(ActionConstant::DELETE)))
+            throw new ForbiddenException(__('Access denied'), new Exception());
+
         return response()->json($this->service->delete($id));
     }
 }
