+++ app/Http/Controllers/SettingController.php
@@ -2,8 +2,13 @@
 
 namespace App\Http\Controllers;
 
+use App\Constants\ActionConstant;
+use App\Constants\PermissionConstant;
+use App\Exceptions\ForbiddenException;
 use App\Services\BaseService;
 use App\Services\SettingService;
+use App\Trait\RoleAndPermissionTrait;
+use Exception;
 use Illuminate\Http\JsonResponse;
 use Illuminate\Http\Request;
 
@@ -11,6 +16,7 @@ class SettingController extends Controller
 {
     public BaseService $service;
 
+    use RoleAndPermissionTrait;
     public function __construct()
     {
         $this->service = new SettingService();
@@ -23,6 +29,9 @@ class SettingController extends Controller
      */
     public function index(): JsonResponse
     {
+        if (!$this->hasPermission(PermissionConstant::setting(ActionConstant::LIST)))
+            throw new ForbiddenException(__('Access denied'), new Exception());
+
         return response()->json($this->service->getAll());
     }
 
@@ -35,6 +44,9 @@ class SettingController extends Controller
      */
     public function store(Request $request): JsonResponse
     {
+        if (!$this->hasPermission(PermissionConstant::setting(ActionConstant::ADD)))
+            throw new ForbiddenException(__('Access denied'), new Exception());
+
         return response()->json($this->service->add($request));
     }
 
@@ -60,6 +72,9 @@ class SettingController extends Controller
      */
     public function update(Request $request, $id): JsonResponse
     {
+        if (!$this->hasPermission(PermissionConstant::setting(ActionConstant::EDIT)))
+            throw new ForbiddenException(__('Access denied'), new Exception());
+
         return response()->json($this->service->update($id,$request));
     }
 
@@ -72,6 +87,9 @@ class SettingController extends Controller
      */
     public function destroy($id): JsonResponse
     {
+        if (!$this->hasPermission(PermissionConstant::setting(ActionConstant::DELETE)))
+            throw new ForbiddenException(__('Access denied'), new Exception());
+
         return response()->json($this->service->delete($id));
     }
 }
