+++ app/Http/Controllers/SpecializeController.php
@@ -2,8 +2,13 @@
 
 namespace App\Http\Controllers;
 
+use App\Constants\ActionConstant;
+use App\Constants\PermissionConstant;
+use App\Exceptions\ForbiddenException;
 use App\Services\BaseService;
 use App\Services\SpecializeService;
+use App\Trait\RoleAndPermissionTrait;
+use Exception;
 use Illuminate\Http\JsonResponse;
 use Illuminate\Http\Request;
 
@@ -11,6 +16,7 @@ class SpecializeController extends Controller
 {
     public BaseService $service;
 
+    use RoleAndPermissionTrait;
     public function __construct()
     {
         $this->service = new SpecializeService();
@@ -23,6 +29,9 @@ class SpecializeController extends Controller
      */
     public function index(): JsonResponse
     {
+        if (!$this->hasPermission(PermissionConstant::specialize(ActionConstant::LIST)))
+            throw new ForbiddenException(__('Access denied'), new Exception());
+
         return response()->json($this->service->getAll());
     }
 
@@ -43,7 +52,7 @@ class SpecializeController extends Controller
      */
     public function create()
     {
-        //
+
     }
 
     /**
@@ -55,6 +64,9 @@ class SpecializeController extends Controller
      */
     public function store(Request $request): JsonResponse
     {
+        if (!$this->hasPermission(PermissionConstant::specialize(ActionConstant::ADD)))
+            throw new ForbiddenException(__('Access denied'), new Exception());
+
         return response()->json($this->service->add($request));
     }
 
@@ -92,6 +104,9 @@ class SpecializeController extends Controller
      */
     public function update(Request $request, $id): JsonResponse
     {
+        if (!$this->hasPermission(PermissionConstant::specialize(ActionConstant::EDIT)))
+            throw new ForbiddenException(__('Access denied'), new Exception());
+
         return response()->json($this->service->update($id, $request));
     }
 
@@ -104,6 +119,9 @@ class SpecializeController extends Controller
      */
     public function destroy($id): JsonResponse
     {
+        if (!$this->hasPermission(PermissionConstant::specialize(ActionConstant::DELETE)))
+            throw new ForbiddenException(__('Access denied'), new Exception());
+
         return response()->json($this->service->delete($id));
     }
 }
